refactor: nest website mutations under an admin.jahia.websites namespace container
The 5 site-lifecycle mutations were grafted directly onto the shared DXM
GqlJahiaAdminMutation (admin.jahia.createSiteByKey, deleteSiteByKey, exportWebsite,
importWebsite, exportAllSites) — generic, unprefixed names with a real cross-module
duplicate-field collision risk on the shared admin namespace.

Move them under one module container: admin.jahia.websites.<op>. New holder type
WebsitesAdminMutation carries the operations as instance @GraphQLField methods;
WebsitesMutation keeps a single 'websites' accessor field (still the registered
extension class). @GraphQLRequiresPermission("websitesAdmin") stays on each leaf.

Also fixes a pre-existing bug: exportWebsite's @GraphQLField/@GraphQLDescription were
orphaned onto the private resolveContainedOrNull helper (left there when that helper was
extracted for a Sonar fix), so exportWebsite was not exposed as a field. Now annotated
correctly. Cypress fixtures/specs updated to the new path.
