+++ tests/cypress/fixtures/graphql/mutation/exportAllSites.graphql
@@ -1,7 +1,9 @@
 mutation ExportAllSites {
     admin {
         jahia {
-            exportAllSites
+            websites {
+                exportAllSites
+            }
         }
     }
 }
