+++ tests/cypress/fixtures/graphql/mutation/exportWebsite.graphql
@@ -1,7 +1,9 @@
 mutation ExportWebsite($siteKey: String!, $exportPath: String!, $onlyStaging: Boolean!) {
     admin {
         jahia {
-            exportWebsite(siteKey: $siteKey, exportPath: $exportPath, onlyStaging: $onlyStaging)
+            websites {
+                exportWebsite(siteKey: $siteKey, exportPath: $exportPath, onlyStaging: $onlyStaging)
+            }
         }
     }
 }
